echo-agent/pyproject.toml
[project]
name = "echo-agent"
version = "0.1.0"
description = "Echo agent for both Daily and LiveKit platforms - responds to ping messages for latency benchmarking with on-demand room creation"
requires-python = ">=3.11"
dependencies = [
    "daily-python>=0.21.0",
    "fastapi>=0.121.2",
    "livekit>=1.0.19",
    "livekit-agents>=1.3.2",
    "pipecat-ai>=0.0.95",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.9.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
